refactor: use global setup for stamps creation in tests
